requirements.txt

fastapi
uvicorn
python-pptx
pdf2image
Pillow
openai
aiofiles
